+++ pex/interpreter_implementation.py
@@ -7,7 +7,7 @@ from pex.enum import Enum
 from pex.typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from typing import FrozenSet, Optional, Tuple
+    from typing import FrozenSet, List, Optional, Tuple
 
 
 class _InterpreterImplementationValue(Enum.Value):
@@ -16,6 +16,7 @@ class _InterpreterImplementationValue(Enum.Value):
         value,  # type: str
         abbr,  # type: str
         binary_name,  # type: str
+        alias=None,  # type: Optional[str]
         free_threaded=None,  # type: Optional[bool]
         initial_version=None,  # type: Optional[Tuple[int, ...]]
     ):
@@ -23,6 +24,7 @@ class _InterpreterImplementationValue(Enum.Value):
         super(_InterpreterImplementationValue, self).__init__(value)
         self.abbr = abbr
         self.binary_name = binary_name
+        self.alias = alias
         self.free_threaded = free_threaded
         self._initial_version = initial_version
 
@@ -52,11 +54,34 @@ class InterpreterImplementation(Enum["InterpreterImplementation.Value"]):
             (cls.CPYTHON_FREE_THREADED, cls.CPYTHON_GIL, cls.PYPY)
         )
 
+    @classmethod
+    def for_value(cls, value):
+        # type: (str) -> InterpreterImplementation.Value
+        for v in cls.values():
+            if v.value == value:
+                return v
+            if v.alias and v.alias == value:
+                return v
+
+        choices = []  # type: List[str]
+        for v in cls.values():
+            choices.append(v.value)
+            if v.alias:
+                choices.append(v.alias)
+        raise ValueError(
+            "{value!r} must be one of {choices}".format(value=value, choices=", ".join(choices))
+        )
+
     CPYTHON = Value("CPython", "cp", "python")
     CPYTHON_FREE_THREADED = Value(
-        "CPython+t", "cp", "python", free_threaded=True, initial_version=(3, 13)
+        "CPython+t",
+        "cp",
+        "python",
+        alias="CPython[free-threaded]",
+        free_threaded=True,
+        initial_version=(3, 13),
     )
-    CPYTHON_GIL = Value("CPython-t", "cp", "python", free_threaded=False)
+    CPYTHON_GIL = Value("CPython-t", "cp", "python", alias="CPython[gil]", free_threaded=False)
     PYPY = Value("PyPy", "pp", "pypy")